Add files via upload
Training advisor
The purpose of this application is to provide career and training advice for software engineering students and professionals.
The application provides both a general learning path available to all users; and a dedicated learning path customized to the user based on his/her progress and proficiency level. The customized learning path is a charged service.
The user (Trainee) will be able to signup for an account, once verified the user can sign in and manage personal and profile data like photo, email, phone…etc. The user can also search for the suitable courses, of the search criteria are ratings and reviews of other users. Once registered in desired course he can solve quizzes and exams, view/ download course materials, and watch videos. The trainee can also view record of his/her progress for the registered course. He may also withdraw from a course. After the training course is completed, a certificate can be issued upon the trainee request after paying necessary fees if applicable.
The Application administration staff are responsible for adding and managing the courses, defining course material and prerequisites. They are also responsible for setting up exams and quizzes. The questions are defined by a third party which maintains a question bank. An online chat based Helpdesk exists to answer trainers, trainees, and visitors’ questions and solve any difficulties or complaints. This is via an online chatting service; this service will be carried out by a third-party component.